requirements.txt
tika
rake-nltk
flask
flask-apscheduler
mysql-connector-python
python-dotenv
